Fix build stuff
I can't seem to get kernels to boot using the regular repack_img method so I'm resorting to using the traditional hammerhead method.
The downside to this is that we're only using gzip for compression instead of LZ4 but it's not too major of a deal IMO. I'll probably try to figure out why the other method isn't working for me in the future.